+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unpublished Changes
+* larva-js:  Adding aria-hidden attribute to tooltips. Closing tooltips on focusout.
+* larva-patterns: Adding aria attributes to c-link and o-account-menu.
 
 ## 1.75.0 12-04-2025
 *  Add conditional check for GA tracking in c-button component